requirements.txt
aiohttp==3.9.0b0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
Deprecated==1.2.14
Django==5.0.7
django-cors-headers==4.4.0
django-ninja==1.2.2
frozenlist==1.4.1
gigachat==0.1.31
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
pillow==10.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyGithub==2.3.0
gunicorn==22.0.0
PyJWT==2.8.0
PyNaCl==1.5.0
pypdf==4.3.1
requests==2.32.3
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
wrapt==1.16.0
yarl==1.9.4
